requirements.txt
streamlit
langgraph
langchain-openai
langchain-community
faiss-cpu
python-dotenv
joblib
scikit-learn
pandas
numpy
